retail/requirements.txt
umpy==1.14.2
pandas==0.22.0
scipy==1.0.1
